+++ edsnlp/pipes/trainable/span_linker/span_linker.py
@@ -154,7 +154,7 @@ class TrainableSpanLinker(
             embedding=eds.span_pooler(
                 hidden_size=128,
                 embedding=eds.transformer(
-                    model="prajjwal1/bert-tiny",
+                    model="hf-internal-testing/tiny-random-bert",
                     window=128,
                     stride=96,
                 ),