+++ changelog.md
@@ -11,6 +11,7 @@
 - Use protocol-provided filesystem when writing parquet files, ie now only telling the program to write parquets to "hdfs:///users/x/y" should suffice.
 - Optimized CRF implementation
 - Fix average precision computation to account for the first recall step.
+- In docs and test, we now use `hf-internal-testing/tiny-random-bert` instead of the outdated `prajjwal1/bert-tiny`
 
 ## v0.20.0 (2025-12-12)